Add files via upload
This is a project using Yelp reviews. I have used python packages to build a model that predicts the number of stars a reviewer will give based on their review.
The model is trained on previous reviews to identify keywords in the reviews that are found in highly starred ratings.